requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cryptography==44.0.2
DateTime==5.0
dj-rest-auth==7.0.1
Django==4.2
django-allauth==65.1.0
django-cors-headers==4.4.0
django-environ==0.12.0
django-modeltranslation==0.18.11
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fiona==1.10.1
GeoAlchemy2==0.17.1
geopandas==0.13.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
numpy==1.24.4
oauthlib==3.2.2
packaging==24.2
pandas==2.0.3
phonenumberslite==9.0.2
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pyproj==3.5.0
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
shapely==2.0.7
six==1.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
zipp==3.20.2
zope.interface==7.2
